Add "About" tab, UI improvements, and debug info
Added a new "About" tab in the `ConfigWindow` class to provide detailed information and instructions on using the HUD Layout Shortcuts plugin. Implemented the `DrawAbout` method to display this information. Replaced the combo box for selecting the redo strategy with radio buttons for better user interaction. Updated the "Save" button text to "Save & Apply Settings" for clarity. Enhanced the `DrawDebugInfo` method to show the current HUD layout index. Added a TODO comment in the `Plugin` class to consider opening the plugin interface directly to the "About" tab in the future.