Add TODOs to OverlayWindow; adjust black_color transparency
Added several TODO comments to the `OverlayWindow` class for future improvements, such as adding toggles for guidelines, automatically opening the window when the HUD Layout editor is open, and renaming the feature to "Alignment helper". Modified the alpha value of the `black_color` variable from 125 to 175 to change the transparency level for showing the corners of other elements.