Add shortcut hints feature and improve config handling
Introduced a new feature to display shortcut hints in the HUD Layout Editor. Added a new configuration option `ShowShortcutHints` to toggle this feature. Created a `ShortcutHintsWindow` class to display the hints and updated the `Plugin` class to manage this window. Enhanced the `Keybind` class with a `ShortText` property and updated the keybindings list accordingly.

Improved configuration handling by adding a `savedConfigHash` field to the `ConfigWindow` class and setting the `ImGuiTabItemFlags.UnsavedDocument` flag for the "Settings" tab if the configuration has changed. Removed the movability logic from the `PreDraw` method in `ConfigWindow`. Updated the `Utils` class to include the `AddonHudLayoutWindow` pointer in the `IsHudLayoutReady` method.